functional_tests/requirements.txt
pytest>=4.6.5
tox>=3.14.0
allure-pytest
paramiko
requests
pytest-timeout
allpairspy